fix: generate token list
